+++ pyproject.toml
@@ -22,7 +22,7 @@ classifiers = [
     'Topic :: Scientific/Engineering :: Artificial Intelligence',
 ]
 dependencies = [
-    "gymnasium>=0.28.1,<0.30",
+    "gymnasium>=0.28.1",
     "numpy >=1.21.0",
     "pygame >=2.1.0",
     "scipy >=1.7.3",
